+++ lib/as2/message.rb
@@ -97,12 +97,17 @@ module As2
       }
     end
 
-    def initialize(message, private_key, public_certificate)
+    def initialize(message, private_key, public_certificate, mic_algorithm: nil)
       # TODO: might need to use OpenSSL::PKCS7.read_smime rather than .new sometimes
       @pkcs7 = OpenSSL::PKCS7.new(message)
       @private_key = private_key
       @public_certificate = public_certificate
       @verification_error = nil
+
+      @mic_algorithm = mic_algorithm || 'sha256'
+      if !As2::DigestSelector.valid?(@mic_algorithm)
+        raise ArgumentError, "'#{@mic_algorithm}' is not a valid MIC algorithm."
+      end
     end
 
     def decrypted_message
@@ -192,7 +197,7 @@ module As2
     end
 
     def mic_algorithm
-      'sha256'
+      @mic_algorithm
     end
 
     # Return the attached file, use .filename and .body on the return value